move constants for mock in own module
