+++ src/mock/destinations.js
@@ -1,32 +1,10 @@
 import {getRandomInteger} from './random.js';
-
-const PHOTOS_COUNT = 5;
-const INFO_COUNT = 5;
-
-const DESTINATIONS = [
-  `Amsterdam`,
-  `Frankfurt`,
-  `Stambul`,
-  `Kiev`,
-  `Paris`,
-  `Montreal`,
-  `Toronto`,
-  `Ottava`
-];
-
-const INFOS = [
-  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-  `Fusce tristique felis at fermentum pharetra.`,
-  `Aliquam id orci ut lectus varius viverra.`,
-  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-  `Aliquam erat volutpat.`,
-  `Nunc fermentum tortor ac porta dapibus.`,
-  `In rutrum ac purus sit amet tempus.`
-];
+import {
+  INFO_COUNT,
+  PHOTOS_COUNT,
+  INFOS,
+  DESTINATIONS
+} from './constants';
 
 const generateDestinationInfo = (infoCount) => new Array(infoCount)
   .fill()
